+++ app/src/main/java/mil/nga/msi/datasource/dgpsstation/DgpsStationMapItem.kt
@@ -3,7 +3,7 @@ package mil.nga.msi.datasource.dgpsstation
 import androidx.room.ColumnInfo
 
 data class DgpsStationMapItem(
-   @ColumnInfo(name = "feature_number") val featureNumber: Int,
+   @ColumnInfo(name = "feature_number") val featureNumber: Float,
    @ColumnInfo(name = "volume_number") val volumeNumber: String,
    @ColumnInfo(name = "latitude") val latitude: Double,
    @ColumnInfo(name = "longitude") val longitude: Double,
